WIP: remove timestamp settings from admin panel
and add a limit/month input
