+++ src/services/Check.php
@@ -221,10 +221,9 @@ class Check
             'state',
             'ts_authority',
             'ts_cert',
+            'ts_limit',
             'ts_login',
-            'ts_override',
             'ts_password',
-            'ts_share',
             'ts_url',
             'title',
             'unreference',
