+++ src/sql/structure.sql
@@ -584,13 +584,6 @@ CREATE TABLE `teams` (
   `link_name` text NOT NULL,
   `link_href` text NOT NULL,
   `datetime` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-  `ts_authority` varchar(255) NOT NULL DEFAULT 'dfn',
-  `ts_login` varchar(255) NULL DEFAULT NULL,
-  `ts_password` varchar(255) NULL DEFAULT NULL,
-  `ts_url` varchar(255) NULL DEFAULT NULL,
-  `ts_cert` varchar(255) NULL DEFAULT NULL,
-  `ts_hash` varchar(10) NOT NULL DEFAULT 'sha256',
-  `ts_override` tinyint(1) UNSIGNED NOT NULL DEFAULT 0,
   `orgid` varchar(255) DEFAULT NULL,
   `public_db` tinyint(1) UNSIGNED NOT NULL DEFAULT 0,
   `force_canread` varchar(255) NOT NULL DEFAULT 'team',
