+++ src/ts/common.ts
@@ -154,29 +154,6 @@ document.addEventListener('DOMContentLoaded', () => {
   }
   listenTrigger();
 
-  /**
-   * Timestamp provider select
-   */
-  if (document.getElementById('ts_authority')) {
-    const select = (document.getElementById('ts_authority') as HTMLSelectElement);
-    const noAccountTsa = ['dfn', 'digicert', 'sectigo', 'globalsign'];
-    select.addEventListener('change', () => {
-      if (noAccountTsa.includes(select.value)) {
-        // mask all
-        document.getElementById('ts_loginpass').toggleAttribute('hidden', true);
-        document.getElementById('ts_urldiv').toggleAttribute('hidden', true);
-      } else if (select.value === 'universign') {
-        // only make loginpass visible
-        document.getElementById('ts_loginpass').removeAttribute('hidden');
-        document.getElementById('ts_urldiv').toggleAttribute('hidden', true);
-      } else if (select.value === 'custom') {
-        // show all
-        document.getElementById('ts_loginpass').removeAttribute('hidden');
-        document.getElementById('ts_urldiv').removeAttribute('hidden');
-      }
-    });
-  }
-
   adjustHiddenState();
 
   /**
